Make ProxyTrack build clean under -Wsign-compare (#1795)
Seven comparisons tested a signed value against a `size_t` or an
`unsigned long`. Four are `send()` and `sendto()` results. The wrap
already took the error branch the code wants, so the cast writes down
what the compiler was doing.

`store.c`'s `fread` site also loses an `(int)` cast on the way into a
`size_t`. `fetchSize` is bounded by an `int` a few lines above, so the
truncation was never reachable. Dropping the cast is what lets the
comparison be typed.

The cast at `proxytrack.c:1291` goes on the `send()` result, not on
`element->size`. That size comes from the cache archive. One whose low
32 bits are all ones would compare equal to a `-1` error, reading a
failed send as a complete one.

Closes #1793

---------

+++ src/proxy/store.c
@@ -988,7 +988,7 @@ int PT_LoadCache__New(PT_Index index_, const char *filename) {
                                          0) == Z_OK) {
               long int pos = (long int) unzGetOffset(zFile);
 
-              assertf(readSizeHeader < sizeof(comment));
+              assertf((size_t) readSizeHeader < sizeof(comment));
               comment[readSizeHeader] = '\0';
               entries++;
               if (pos > 0) {
@@ -1300,7 +1300,9 @@ static PT_Element PT_ReadCache__New_u(PT_Index index_, const char *url,
                 if (flags & FETCH_BODY) {
                   r->adr = (char *) malloc(r->size + 1);
                   if (r->adr != NULL) {
-                    if (unzReadCurrentFile(index->zFile, r->adr, (unsigned int) r->size) != r->size) {  // erreur
+                    if ((size_t) unzReadCurrentFile(index->zFile, r->adr,
+                                                    (unsigned int) r->size) !=
+                        r->size) {
                       PT_Element_DropBody(r);
                       r->statuscode = STATUSCODE_INVALID;
                       strcpybuff(r->msg, "Cache Read Error : Read Data");
@@ -2647,10 +2649,9 @@ static PT_Element PT_ReadCache__Arc_u(PT_Index index_, const char *url,
               if (r->statuscode != STATUSCODE_INVALID) {
                 r->adr = (char *) malloc(fetchSize);
                 if (r->adr != NULL) {
-                  if (fetchSize > 0
-                      && (r->size =
-                          (int) fread(r->adr, 1, fetchSize,
-                                      index->file)) != fetchSize) {
+                  if (fetchSize > 0 &&
+                      (r->size = fread(r->adr, 1, fetchSize, index->file)) !=
+                          (size_t) fetchSize) {
                     int last_errno = errno;
 
                     r->statuscode = STATUSCODE_INVALID;
